backend/requirements.txt
# Amal Backend Requirements
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0

# ML Dependencies (from intent_model and rag_scientific)
torch>=2.0.0
transformers>=4.35.0
joblib>=1.3.0
numpy>=1.24.0

# RAG Dependencies
chromadb>=0.4.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
